Crate fedimint_build

Fedimint build scripts

Implements detection of build hash. Used both internally in all public-consumed binaries of fedimint, and also in custom builds that can included 3rd party Fedimint modules.

To use include:

[build-dependencies]
fedimint-build = { version = "=0.4.0-alpha", path = "../fedimint-build" }

in Cargo.toml, and:

fn main() {
    fedimint_build::set_code_version();
}

in build.rs script.

This will define FEDIMINT_BUILD_CODE_VERSION at the build time, which can be accessed via fedimint_build_code_version_env!() and passed to binary builders like FedimintCli::new.

Modules

envs

Functions

call_cmd 🔒

set_code_version

Run from a build.rs script to detect code version. See crate for description.

set_code_version_inner 🔒