fedimint_build

Module envs

Constants

FEDIMINT_BUILD_CODE_VERSION_ENV

Env variable the cargo will set during crate build to pass the detected git hash to the binary itself.

FORCE_GIT_HASH_ENV

Env variable to set to force git hash during build process